feat() Finally figure out how to add pagination
Note: the jekyll documentation is not very nice. It wasn't very clear right away what the pagination_path is doing and where to put the pagination-render code
Do not change index.html in the root folder! It's what the pagination is taking as its default file that makes the pagination possible for some magical reason